Add account

+++ server/src/data/persister/sqlite/todo.ts
@@ -1,5 +1,3 @@
-import { Database, OPEN_CREATE, OPEN_READWRITE } from "sqlite3";
-import { DatabaseConnectionError } from "../../error/database_connection";
 import { DatabaseQueryError } from "../../error/database_query_error";
 import { HTTPNotFoundError } from "../../error/http_404";
 import { TodoPayload } from "../../payload/todo";
@@ -7,26 +5,10 @@ import { SerializableTodo } from "../../serializable_todo";
 import { Todo } from "../../todo";
 import { UUIDv4 } from "../../uuidv4";
 import { TodoPersister } from "../todo";
+import { SQLite3AccountPersister } from "./account";
+import { SQLite3Persister } from "./sqlite";
 
-export class SQLite3TodoPersister implements TodoPersister {
-    protected readonly path: string;
-    protected db: Database;
-
-    public constructor(path: string) {
-        this.path = path;
-    }
-
-    protected async connect(): Promise<void> {
-        return new Promise((resolve, reject) => {
-            this.db = new Database(this.path, OPEN_READWRITE | OPEN_CREATE, (error: Error | null) => {
-                if (error !== null) {
-                    reject(new DatabaseConnectionError(`No databse connection (${this.path})`));
-                }
-                resolve()
-            });
-        })
-    }
-
+export class SQLite3TodoPersister extends SQLite3Persister implements TodoPersister {
     protected async create_tables(): Promise<void> {
         return new Promise((resolve, reject) => {
             this.db.run(`
@@ -34,7 +16,9 @@ export class SQLite3TodoPersister implements TodoPersister {
                     id text not null check(length(id) == 36) primary key,
                     title text not null,
                     description text not null,
-                    state int check(state <= 1)
+                    state int check(state <= 1),
+                    accountid text not null,
+                    foreign key(accountid) references account(id)
                 );
             `, (error: Error | null) => {
                 if (error !== null) {
@@ -45,19 +29,10 @@ export class SQLite3TodoPersister implements TodoPersister {
         });
     }
 
-    protected async get_db(): Promise<Database> {
-        if (this.db !== undefined) {
-            return this.db;
-        }
-        await this.connect();
-        await this.create_tables();
-        return this.db;
-    }
-
-    public async get_todo(id: UUIDv4): Promise<Todo> {
+    public async get_todo(id: UUIDv4, user_id: UUIDv4): Promise<Todo> {
         return new Promise(async (resolve, reject) => {
             const db = await this.get_db();
-            db.get("select * from todo where id = ?;", [id.toString()], (error: Error | null, todo: Todo | undefined) => {
+            db.get("select * from todo where id = ? and accountid = ?;", [id, user_id], (error: Error | null, todo: Todo | undefined) => {
                 if (error !== null) {
                     return reject(error)
                 }
@@ -69,10 +44,10 @@ export class SQLite3TodoPersister implements TodoPersister {
         });
     }
 
-    public async get_todos(): Promise<Todo[]> {
+    public async get_todos(user_id: UUIDv4): Promise<Todo[]> {
         return new Promise(async (resolve, reject) => {
             const db = await this.get_db();
-            db.all("select * from todo;", (error: Error | null, todos: Todo[]) => {
+            db.all("select * from todo where accountid = ?;", [user_id], (error: Error | null, todos: Todo[]) => {
                 if (error !== null) {
                     return reject(error);
                 }
@@ -84,11 +59,11 @@ export class SQLite3TodoPersister implements TodoPersister {
         });
     }
 
-    public async create_todo(todo: TodoPayload): Promise<Todo> {
+    public async create_todo(todo: TodoPayload, user_id: UUIDv4): Promise<Todo> {
         return new Promise(async (resolve, reject) => {
             const db = await this.get_db();
             const id = UUIDv4.generate();
-            db.run("insert into todo values (?, ?, ?, ?)", [id, todo.title, todo.description, todo.state], (error: Error | null) => {
+            db.run("insert into todo values (?, ?, ?, ?, ?)", [id, todo.title, todo.description, todo.state, user_id], (error: Error | null) => {
                 if (error !== null) {
                     return reject(error);
                 }
@@ -97,10 +72,10 @@ export class SQLite3TodoPersister implements TodoPersister {
         });
     }
 
-    public async mutate_todo(todo: Todo): Promise<Todo> {
+    public async mutate_todo(todo: Todo, user_id: UUIDv4): Promise<Todo> {
         return new Promise(async (resolve, reject) => {
             const db = await this.get_db();
-            db.run("update todo set title = ?, description = ?, state = ? where id = ?", [todo.title, todo.description, todo.state, todo.id], (error: Error | null) => {
+            db.run("update todo set title = ?, description = ?, state = ? where id = ? and accountid = ?", [todo.title, todo.description, todo.state, todo.id, user_id], (error: Error | null) => {
                 if (error !== null) {
                     return reject(error);
                 }
@@ -109,10 +84,10 @@ export class SQLite3TodoPersister implements TodoPersister {
         });
     }
 
-    public async delete_todo(id: UUIDv4): Promise<void> {
+    public async delete_todo(id: UUIDv4, user_id: UUIDv4): Promise<void> {
         return new Promise(async (resolve, reject) => {
             const db = await this.get_db();
-            db.run("delete from todo where id = ?", [id], (error: Error | null) => {
+            db.run("delete from todo where id = ? where accountid = ?", [id, user_id], (error: Error | null) => {
                 if (error !== null) {
                     return reject(error);
                 }
